Add worker
