+++ app/src/main/java/ru/studyguk/cryptoapp/data/repository/CoinRepositoryImpl.kt
@@ -3,17 +3,17 @@ package ru.studyguk.cryptoapp.data.repository
 import android.app.Application
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.Transformations
-import kotlinx.coroutines.delay
+import androidx.work.ExistingWorkPolicy
+import androidx.work.WorkManager
 import ru.studyguk.cryptoapp.data.database.AppDatabase
 import ru.studyguk.cryptoapp.data.mapper.CoinMapper
-import ru.studyguk.cryptoapp.data.network.ApiFactory
+import ru.studyguk.cryptoapp.data.workers.RefreshDataWorker
 import ru.studyguk.cryptoapp.domain.CoinInfo
 import ru.studyguk.cryptoapp.domain.CoinRepository
 
 class CoinRepositoryImpl(private val application: Application) : CoinRepository {
 
     private val coinInfoDao = AppDatabase.getInstance(application).coinPriceInfoDao()
-    private val apiService = ApiFactory.apiService
 
     private val mapper = CoinMapper()
 
@@ -31,18 +31,12 @@ class CoinRepositoryImpl(private val application: Application) : CoinRepository 
         }
     }
 
-    override suspend fun loadData() {
-        while (true) {
-            try {
-                val topCoins = apiService.getTopCoinsInfo()
-                val fSyms = mapper.mapNamesListToString(topCoins)
-                val jsonContainer = apiService.getFullPriceList(fSyms = fSyms)
-                val coinInfoDtoList = mapper.mapJsonContainerToListCoinInfo(jsonContainer)
-                val dbModelList = coinInfoDtoList.map { mapper.mapDtoToDbModel(it) }
-                coinInfoDao.insertPriceList(dbModelList)
-            } catch (e: Exception) {
-            }
-            delay(10000)
-        }
+    override fun loadData() {
+        val workManager = WorkManager.getInstance(application)
+        workManager.enqueueUniqueWork(
+            RefreshDataWorker.NAME,
+            ExistingWorkPolicy.REPLACE,
+            RefreshDataWorker.makeRequest()
+        )
     }
 }
