+++ app/src/main/java/ru/studyguk/cryptoapp/domain/LoadDataUseCase.kt
@@ -2,5 +2,5 @@ package ru.studyguk.cryptoapp.domain
 
 class LoadDataUseCase(private val repository: CoinRepository) {
 
-    suspend operator fun invoke() = repository.loadData()
+    operator fun invoke() = repository.loadData()
 }
