+++ app/src/main/java/ru/studyguk/cryptoapp/presentation/CoinViewModel.kt
@@ -20,9 +20,7 @@ class CoinViewModel(application: Application) : AndroidViewModel(application) {
     val coinInfoList = getCoinInfoListUseCase()
 
     init {
-        viewModelScope.launch {
-            loadDataUseCase()
-        }
+        loadDataUseCase()
     }
 
     fun getDetailInfo(fSym: String) = getCoinInfoUseCase(fSym)
